리뷰 보기 통신 완료

+++ app/src/main/java/dmzing/workd/model/review/reviewDto.kt
@@ -6,7 +6,7 @@ data class reviewDto(
         var title : String,
         var thumbnailUrl : String,
         var courseId : Int,
-        var startAt : Int,
-        var endAt : Int,
+        var startAt : Long,
+        var endAt : Long,
         var postDto : ArrayList<PostDto>
 )
